+++ ecdsa/Cargo.toml
@@ -14,6 +14,7 @@ categories = ["cryptography", "no-std"]
 keywords   = ["crypto", "ecc", "nist", "secp256k1", "signature"]
 
 [dependencies]
+der = { version = "=0.2.0-pre", optional = true, features = ["big-uint"] }
 elliptic-curve = { version = "=0.9.0-pre", default-features = false }
 hmac = { version = "0.10", optional = true, default-features = false }
 signature = { version = ">= 1.3.0, < 1.4.0", default-features = false, features = ["rand-preview"] }
@@ -30,7 +31,7 @@ arithmetic = ["elliptic-curve/arithmetic"]
 dev = ["arithmetic", "digest", "zeroize"]
 digest = ["elliptic-curve/digest", "signature/digest-preview"]
 hazmat = []
-pkcs8 = ["elliptic-curve/pkcs8"]
+pkcs8 = ["elliptic-curve/pkcs8", "der"]
 pem = ["elliptic-curve/pem", "pkcs8"]
 sign = ["arithmetic", "digest", "hazmat", "hmac", "zeroize"]
 std = ["alloc", "elliptic-curve/std", "signature/std"]
